Pedro

+++ CRM-BACKEND/src/main/java/solution_market_CRM/repository/ClientRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import solution_market_CRM.model.Client;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long>
+public interface ClientRepository extends JpaRepository<Client, Integer>
 {
 
 }
