+++ CRM-BACKEND/src/main/java/solution_market_CRM/repository/EmployeeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import solution_market_CRM.model.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long>
+public interface EmployeeRepository extends JpaRepository<Employee, Integer>
 {
     
 }
